extract sortEntities in its own file
